ecommerce-PQ/requirements.txt
azure-functions
pyodbc
azure-identity
azure-keyvault-secrets
langchain
langchain-community
langchain-experimental
langchain-openai
